premier exo

+++ basics/src/main/java/be/ecam/basics/exercises/Counter.java
@@ -1,8 +1,8 @@
 package be.ecam.basics.exercises;
 
 public class Counter {
-    public static byte count(int start, int steps) {
-        byte c = (byte) start;
+    public static long count(int start, int steps) {
+        long c = (long) start;
         for (int i = 0; i < steps; i++) {
             c++;
         }
